# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Essential HICS API
  description: |
    Hospital Incident Command System — Phase 1 REST API.
    Covers authentication, RBAC, organizational structure, and audit logging.
  version: 1.0.0
  contact:
    name: HICS Support
    email: support@apexhealth.example
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.hics.yourhospital.org
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HicsRole:
      type: string
      enum:
        - SYSTEM_ADMIN
        - SYSTEM_VIEWER
        - FACILITY_ADMIN
        - INCIDENT_COMMANDER
        - DEPUTY_INCIDENT_COMMANDER
        - PUBLIC_INFORMATION_OFFICER
        - SAFETY_OFFICER
        - LIAISON_OFFICER
        - MEDICAL_TECHNICAL_SPECIALIST
        - OPERATIONS_SECTION_CHIEF
        - PLANNING_SECTION_CHIEF
        - LOGISTICS_SECTION_CHIEF
        - FINANCE_ADMIN_SECTION_CHIEF
        - MEDICAL_CARE_BRANCH_DIRECTOR
        - INFRASTRUCTURE_BRANCH_DIRECTOR
        - SECURITY_BRANCH_DIRECTOR
        - BUSINESS_CONTINUITY_BRANCH_DIRECTOR
        - HAZMAT_BRANCH_DIRECTOR
        - PATIENT_FAMILY_INFORMATION_BRANCH_DIRECTOR
        - RESOURCES_UNIT_LEADER
        - SITUATION_UNIT_LEADER
        - DOCUMENTATION_UNIT_LEADER
        - DEMOBILIZATION_UNIT_LEADER
        - SERVICE_BRANCH_DIRECTOR
        - SUPPORT_BRANCH_DIRECTOR
        - COMMUNICATIONS_UNIT_LEADER
        - IT_SYSTEMS_UNIT_LEADER
        - STAFF_FOOD_WATER_UNIT_LEADER
        - TRANSPORTATION_UNIT_LEADER
        - FACILITIES_UNIT_LEADER
        - SUPPLY_UNIT_LEADER
        - LABOR_POOL_CREDENTIALS_UNIT_LEADER
        - TIME_UNIT_LEADER
        - PROCUREMENT_UNIT_LEADER
        - COMPENSATION_CLAIMS_UNIT_LEADER
        - COST_UNIT_LEADER
        - RESPONDER
        - READ_ONLY_OBSERVER
    Facility:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        shortName: { type: string }
        address:
          type: object
          properties:
            street: { type: string }
            city: { type: string }
            state: { type: string }
            zip: { type: string }
        phone: { type: string }
        fax: { type: string }
        licenseNumber: { type: string }
        facilityType: { type: string, enum: [HOSPITAL, CLINIC, ALTERNATE_CARE_SITE, OTHER] }
        isActive: { type: boolean }
        timezone: { type: string, example: "America/New_York" }
        emergencyContactName: { type: string }
        emergencyContactPhone: { type: string }
        createdAt: { type: string, format: date-time }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
security:
  - bearerAuth: []
paths:
...
  /api/facilities:
    get:
      tags: [Facilities]
      summary: List all facilities
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Facility' }

    post:
      tags: [Facilities]
      summary: Create a new facility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, shortName, address]
              properties:
                name: { type: string }
                shortName: { type: string }
                address:
                  type: object
                  required: [street, city, state, zip]
                  properties:
                    street: { type: string }
                    city: { type: string }
                    state: { type: string, minLength: 2, maxLength: 2 }
                    zip: { type: string }
                phone: { type: string }
                fax: { type: string }
                facilityType: { type: string, enum: [HOSPITAL, CLINIC, ALTERNATE_CARE_SITE, OTHER] }
                timezone: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facility' }
  /api/facilities/{id}:
    get:
      tags: [Facilities]
      summary: Get facility by ID (includes departments)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facility' }

    put:
      tags: [Facilities]
      summary: Update facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facility' }
  /api/facilities/{id}/departments:
    get:
      tags: [Facilities]
      summary: List departments for a facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    name: { type: string }
                    code: { type: string }
                    parentDepartmentId: { type: string, format: uuid }

    post:
      tags: [Facilities]
      summary: Create a department
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name: { type: string }
                code: { type: string }
                parentDepartmentId: { type: string, format: uuid }
      responses:
        '201': { description: Department created }
  /api/facilities/{id}/users:
    get:
      tags: [Facilities]
      summary: List users assigned to a facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 25 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array }
                  pagination: { $ref: '#/components/schemas/Pagination' }

    post:
      tags: [Facilities]
      summary: Create and add a user to this facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, firstName, lastName, hicsRole]
              properties:
                email: { type: string, format: email }
                firstName: { type: string }
                lastName: { type: string }
                jobTitle: { type: string }
                employeeId: { type: string }
                phoneMobile: { type: string }
                hicsRole: { $ref: '#/components/schemas/HicsRole' }
      responses:
        '201': { description: User created and invite email sent }
  /api/facilities/{id}/users/import:
    post:
      tags: [Facilities]
      summary: Bulk import users from CSV
      description: |
        Validates all rows before importing any. Returns row-by-row errors on failure.
        CSV columns: first_name, last_name, email, job_title, employee_id, phone_mobile, hics_role
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: integer }
                  skipped: { type: integer }
                  skippedEmails: { type: array, items: { type: string } }
        '422':
          description: Validation failed — no users were imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        row: { type: integer }
                        errors: { type: array, items: { type: string } }
  /api/facilities/{id}/positions:
    get:
      tags: [Facilities]
      summary: Get HICS org chart positions for a facility
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    hicsRole: { $ref: '#/components/schemas/HicsRole' }
                    displayName: { type: string }
                    reportsToRole: { $ref: '#/components/schemas/HicsRole' }

  # ── Audit Log ──────────────────────────────────────────────────────────────────
...